Database types updated
