+++ frontend/src/metabase/status/components/DatabaseStatusSmall/DatabaseStatusSmall.styled.tsx
@@ -2,10 +2,10 @@ import styled from "@emotion/styled";
 import { color, lighten } from "metabase/lib/colors";
 import Icon from "metabase/components/Icon";
 import LoadingSpinner from "metabase/components/LoadingSpinner";
-import { InitialSyncStatus } from "metabase-types/api";
+import { DatabaseInitialSyncStatus } from "metabase-types/api";
 
 interface Props {
-  status: InitialSyncStatus;
+  status: DatabaseInitialSyncStatus;
 }
 
 const getIconColor = ({ status }: Props): string => {
